fix(croquis-cf): split fallthrough attr helpers
